+++ Tests/LinuxMain.swift
@@ -1,6 +1,6 @@
-import XCTest
 @testable import Finite
 import FiniteTests
+import XCTest
 
 var tests = [XCTestCaseEntry]()
 tests += FiniteTests.allTests()